refactor: integrate domain and entity models and restructure project directory layout
